server/requirements.txt
flask==3.0.0
flask-socketio==5.3.6
flask-cors==4.0.0
python-socketio==5.10.0
eventlet==0.33.3
pillow==10.1.0
python-magic-bin==0.4.14
